domains_controller admin_delete find count code to ensure the proper permission to delete this domain by the right people
this code to be used for edit/delete/make_cover  products, product_images, payments,
shipping_rates, themes switching, blogs, posts, webpages,etc in future

domains/admin_index to display delete link only for non shop_web_address domains

staging copy database add another column default false for shop_web_address in domains

mainsite Merchant model ensure domain is set true for shop_web_address during registration
mainsite register view display the correct main domain by altering the view and
action in controller

+++ app/controllers/domains_controller.php
@@ -80,16 +80,37 @@ class DomainsController extends AppController {
 		$this->set(compact('shops'));
 	}
 
-	function admin_delete($id = null) {
+	function admin_delete($id = false) {
 		if (!$id) {
 			$this->Session->setFlash(__('Invalid id for domain', true), 'default', array('class'=>'flash_failure'));
 			$this->redirect(array('action'=>'index'));
 		}
-		if ($this->Domain->delete($id)) {
-			$this->Session->setFlash(__('Domain deleted', true), 'default', array('class'=>'flash_success'));
-			$this->redirect(array('action'=>'index'));
+		
+		$userId = User::get('User.id');
+		$shopId = Shop::get('Shop.id');
+		$this->Domain->Shop->Merchant->Behaviors->attach('Linkable.Linkable');
+		
+		$merchantAllowed = $this->Domain->Shop->Merchant->find('count',
+								       array('conditions' =>
+									     array('Merchant.user_id' => $userId,
+										   'Merchant.shop_id' => $shopId,
+										   'Domain.id'=>$id,
+										   'Domain.shop_web_address' => false),
+									     'link' =>
+									     array('Shop' => array('Domain'))
+									     ));
+		
+		
+		if ($merchantAllowed) {
+			if ($this->Domain->delete($id)) {
+				$this->Session->setFlash(__('Domain deleted', true), 'default', array('class'=>'flash_success'));
+				$this->redirect(array('action'=>'index'));
+			}
+			$this->Session->setFlash(__('Domain was not deleted', true), 'default', array('class'=>'flash_failure'));
+		} else {
+			$this->Session->setFlash(__('You are not authorized to delete this domain', true), 'default', array('class'=>'flash_failure'));
 		}
-		$this->Session->setFlash(__('Domain was not deleted', true), 'default', array('class'=>'flash_failure'));
+		
 		$this->redirect(array('action' => 'index'));
 	}
